Final_Project/requirements.txt
python-dotenv==1.0.0
Requests==2.31.0
web3==6.2.0
